* src/engine/Account.c: move currency conversion to gnc-pricedb.c
        * src/engine/gnc-pricedb.c
        * src/engine/gnc-pricedb.h: add functions
	  gnc_pricedb_convert_balance_latest_price and
          gnc_pricedb_convert_balance_nearest_price

        * src/engine/gw-engine-spec.scm: wrap functions
          gnc_price_list_destroy, gnc_pricedb_lookup_latest_any_currency,
          gnc_pricedb_lookup_nearest_in_time_any_currency,
          gnc_pricedb_convert_balance_latest_price and
          gnc_pricedb_convert_balance_nearest_price

        * src/report/report-system/commodity-utilities.scm: use
          gnc:pricedb-convert-balance-latest-price and
          gnc:pricedb-convert-balance-nearest-price to improve
          multiple currency reports

        * src/report/standard-reports/advanced-portfolio.scm:
          improve report for stocks and funds not denominated in report currency


git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/branches/1.8@8256 57a11ea4-9604-0410-9ed3-97b8803252fd

+++ src/import-export/hbci/hbci-interaction.c
@@ -252,7 +252,7 @@ static int msgInputPin(const HBCI_User *user,
     if (!retval)
       break;
     
-    if (strlen(passwd) < minsize) {
+    if (strlen(passwd) < (unsigned int)minsize) {
       gboolean retval;
       char *msg = 
 	g_strdup_printf (  _("The PIN needs to be at least %d characters \n"
